+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/icon/IconsTest.java
@@ -28,7 +28,7 @@ public final class IconsTest {
 
 	@Test
 	void test() {
-		Icons icons = Icons.icons();
+		Icons icons = Icons.icons(16);
 		assertThrows(IllegalArgumentException.class, () -> icons.get(Foundation.ALERT));
 		icons.add(Foundation.ALERT, Foundation.FOUNDATION);
 		assertThrows(IllegalArgumentException.class, () -> icons.add(Foundation.FOUNDATION));